Disabled the algorithm description of the eqsat python rewriter first
